+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2019-06-18
+
+### Changed
+- Add classes to templates array items nav, item, and item_current in default options.
+
+### Added
+- Classes array indexes for items in the templates array: nav, list, list_item, item, and item_current.
+
 ## [0.2.0] - 2019-06-17
 
 ### Changed